requirements.txt
aiofiles==0.5.0
altair==4.1.0
aniso8601==8.1.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
async-generator==1.10
asyncpg==0.21.0
asynctest==0.13.0
attrs==19.3.0
Authlib==0.15.2
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.0.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.2.1
blessings==1.7
blinker==1.4
boto3==1.15.16
botocore==1.18.16
bottle==0.12.18
bottle-websocket==0.2.9
bpython==0.19
Brlapi==0.7.0
bs4==0.0.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorclass==2.2.0
colorlog==4.6.2
command-not-found==0.3
cryptography==2.9.2
cupshelpers==1.0
curtsies==0.3.4
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.3
dnspython==2.0.0
docopt==0.6.2
duplicity==0.8.12.0
Eel==0.12.4
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.0.1
exchangelib==3.3.2
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
fpdf==1.7.2
furl==2.1.0
future==0.18.2
gevent==20.6.0
gevent-websocket==0.10.1
graphviz==0.13.2
greenlet==0.4.16
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.18
httplib2==0.14.0
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.2.1
meld==3.20.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mss==6.1.0
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
netifaces==0.10.4
netsuitesdk==1.12.0
notebook==6.1.4
notifiers==1.2.1
ntlm-auth==1.5.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
openpyxl==3.0.5
orderedmultidict==1.0.1
packaging==20.4
panda3d==1.10.6
pandas==1.1.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
pathtools==0.1.2
pdfminer.six==20201018
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.0.1
pip-upgrader==1.4.15
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyarrow==1.0.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydeck==0.5.0b1
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynput-robocorp-fork==1.7.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyRFC3339==1.1
pyrsistent==0.17.3
pytesseract==0.3.7
pytest==6.0.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xlib==0.29
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.4.0
rfc3987==1.3.8
robotframework==3.2.2
robotframework-pythonlibcore==2.1.0
robotframework-requests==0.6.6
robotframework-seleniumlibrary==4.5.0
robotframework-seleniumtestability==1.1.0
rpaframework==6.7.1
rpaframework-core==3.0.0
s3transfer==0.3.3
sanic==20.6.3
Sanic-Auth==0.2.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simple-salesforce==1.10.1
simplejson==3.16.0
six==1.14.0
sniffio==1.1.0
sortedcontainers==2.3.0
soupsieve==2.0.1
sqlparse==0.4.1
streamlit==0.68.1
systemd-python==234
tabula-py==2.2.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.53.0
traitlets==5.0.4
tweepy==3.9.0
tzdata==2020.4
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==3.1.0
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvloop==0.14.0
validators==0.18.1
virtualenv==20.0.17
wadllib==1.3.3
watchdog==0.10.3
wcwidth==0.2.5
webdrivermanager==0.9.0
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zeep==4.0.0
zipp==1.0.0
zope.event==4.4
zope.interface==5.1.0
